+++ checksum-Elixir.FinanceRustler.Solver.exs
@@ -1,4 +1,5 @@
 %{
-  "libfinance_rustler-v0.2.0-nif-2.15-aarch64-unknown-linux-gnu.so.tar.gz" => "sha256:74c3dc30aa65a078db09fbfc5c117fc953ccb6834253955cf1ce5f033317cfd7",
-  "libfinance_rustler-v0.2.0-nif-2.15-x86_64-unknown-linux-gnu.so.tar.gz" => "sha256:1b25f6d1a7f6cb773d932172077778aefd745ba2e348221bac96d28427ae2ef6",
+  "libfinance_rustler-v0.2.0-nif-2.15-aarch64-apple-darwin.so.tar.gz" => "sha256:6ff559a199d94597a9ec1c95b859c241eb7cea3910997f90914234976038060e",
+  "libfinance_rustler-v0.2.0-nif-2.15-aarch64-unknown-linux-gnu.so.tar.gz" => "sha256:cdcde2cb93151e900bebc1dd591a2427f438929259861ad1becd2fcf8b086bdf",
+  "libfinance_rustler-v0.2.0-nif-2.15-x86_64-unknown-linux-gnu.so.tar.gz" => "sha256:64fc6387942a33f6bad9ade22f105c4919f5c2407e74ceec462f83325032253f",
 }