+++ lib/finance_rustler/solver.ex
@@ -37,7 +37,6 @@ defmodule FinanceRustler.Solver do
     targets: ~w(
       x86_64-unknown-linux-gnu
       aarch64-unknown-linux-gnu
-      x86_64-apple-darwin
       aarch64-apple-darwin
     )